pyproject.toml
[tool.poetry]
name = "shop-of-electronic"
version = "0.1.0"
description = ""
authors = ["Sandi_Navoiy <santes.au@gmail.com>"]
readme = "README.md"
packages = [{include = "shop_of_electronic"}]

[tool.poetry.dependencies]
python = "^3.11"
isort = "^5.12.0"
black = "^23.10.1"
django = "^4.2.6"
djangorestframework = "^3.14.0"
flake8 = "^6.1.0"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.0"
python-dotenv = "^1.0.0"
psycopg2 = "^2.9.9"
drf-spectacular = "^0.26.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
